+++ src/main/java/de/cowtipper/cowlection/chesttracker/data/LowestBinsCache.java
@@ -2,7 +2,7 @@ package de.cowtipper.cowlection.chesttracker.data;
 
 import java.util.HashMap;
 
-public class LowestBinsCache extends HashMap<String, Integer> {
+public class LowestBinsCache extends HashMap<String, Long> {
     public boolean hasData() {
         return size() > 0;
     }
